+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/CurrentDatePicker/CurrentDatePicker.tsx
@@ -1,6 +1,5 @@
-import { useMemo } from "react";
 import { t } from "ttag";
-import { Button, Group, Stack, Select, Tooltip } from "metabase/ui";
+import { Button, Group, Stack, Tooltip } from "metabase/ui";
 import * as Lib from "metabase-lib";
 import type {
   DatePickerTruncationUnit,
@@ -45,34 +44,4 @@ export function CurrentDatePicker({ value, onChange }: CurrentDatePickerProps) {
       ))}
     </Stack>
   );
-}
-
-interface SimpleCurrentDatePickerProps {
-  value: RelativeDatePickerValue;
-  onChange: (value: RelativeDatePickerValue) => void;
-}
-
-export function SimpleCurrentDatePicker({
-  value,
-  onChange,
-}: SimpleCurrentDatePickerProps) {
-  const options = useMemo(
-    () =>
-      UNIT_GROUPS.flatMap(group =>
-        group.map(unit => ({
-          value: unit,
-          label: Lib.describeTemporalUnit(unit),
-        })),
-      ),
-    [],
-  );
-
-  const handleChange = (unitValue: string | null) => {
-    const option = options.find(option => option.value === unitValue);
-    if (option) {
-      onChange({ ...value, unit: option.value });
-    }
-  };
-
-  return <Select value={value.unit} data={options} onChange={handleChange} />;
 }
