+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/CurrentDatePicker/SimpleCurrentDatePicker/index.ts
@@ -1,2 +1,1 @@
-export * from "./CurrentDatePicker";
 export * from "./SimpleCurrentDatePicker";
