+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/DateIntervalPicker/SimpleDateIntervalPicker/index.ts
@@ -1,2 +1,1 @@
-export * from "./DateIntervalPicker";
 export * from "./SimpleDateIntervalPicker";
