+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/RelativeDatePicker.tsx
@@ -9,14 +9,8 @@ import {
   isOffsetIntervalValue,
   setDirection,
 } from "./utils";
-import {
-  CurrentDatePicker,
-  SimpleCurrentDatePicker,
-} from "./CurrentDatePicker";
-import {
-  DateIntervalPicker,
-  SimpleDateIntervalPicker,
-} from "./DateIntervalPicker";
+import { CurrentDatePicker } from "./CurrentDatePicker";
+import { DateIntervalPicker } from "./DateIntervalPicker";
 import { DateOffsetIntervalPicker } from "./DateOffsetIntervalPicker";
 import { TabList } from "./RelativeDatePicker.styled";
 
@@ -85,21 +79,5 @@ export function RelativeDatePicker({
         </Tabs.Panel>
       ))}
     </Tabs>
-  );
-}
-
-interface SimpleRelativeDatePickerProps {
-  value: RelativeDatePickerValue;
-  onChange: (value: RelativeDatePickerValue) => void;
-}
-
-export function SimpleRelativeDatePicker({
-  value,
-  onChange,
-}: SimpleRelativeDatePickerProps) {
-  return isIntervalValue(value) ? (
-    <SimpleDateIntervalPicker value={value} onChange={onChange} />
-  ) : (
-    <SimpleCurrentDatePicker value={value} onChange={onChange} />
   );
 }
