+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/SimpleRelativeDatePicker/index.ts
@@ -1,2 +1,1 @@
-export * from "./RelativeDatePicker";
 export * from "./SimpleRelativeDatePicker";
